feat: useSetPositions 커스텀훅 추가
